refactor(react): rename host.ts -> runtime.ts, /state subpath -> /runtime
The shared module owns the Runtime seam and the render/activation machinery -
it is the runtime, and the name now says so (also avoiding collision with
mvc's jsx-runtime `host()`/`Host` concept). The published subpath follows:
`@expressive/react/state` -> `@expressive/react/runtime`, /state removed.
Preact's import + tsconfig path updated to match.
