+++ packages/react/src/state.get.ts
@@ -1,5 +1,5 @@
 import { State, Context, observer, watch } from '@expressive/mvc';
-import { Runtime, useFactory, useHook, useReady } from './host';
+import { Runtime, useFactory, useHook, useReady } from './runtime';
 
 /** Type may not be undefined - instead will be null.  */
 type NoVoid<T> = T extends undefined | void ? null : T;
